Adding cohort dash feature.
