+++ spec/factories/adjustments.rb
@@ -2,5 +2,4 @@ FactoryGirl.define do
   factory :adjustment do
     checkin
   end
-
 end
