+++ spec/factories/instructors.rb
@@ -1,4 +1,5 @@
 FactoryGirl.define do
   factory :instructor do
+    user
   end
 end
